# api/doc/swagger_api_doc.json
{
  "info": {
    "title": "SHOPING API",
    "description": "This is a sample Petstore server. You can find \nout more about Swagger at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "thonguyen140294@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "swagger": "2.0",
  "host": "bestsporthome-api.herokuapp.com",
  "basePath": "/",
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "get list product",
        "operationId": "getProducts",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/List_Product_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
...
    "/category": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "get category",
        "operationId": "getCategory",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category_Response"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
    "/category/{category_id}/brand": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "get brand of category",
        "operationId": "getBrandOfCategory",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          },
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Brand_Response"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
    "/brand": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "get brand",
        "operationId": "getBrand",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Brand_Response"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
...
  },
  "schemes": [
    "https",
    "http"
  ],
  "definitions": {
...
    "Error_Response": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": [
        "message"
      ]
    },
...
    "Product_Response": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "category": {
          "type": "string"
        },
        "sale": {
          "type": "integer"
        },
        "image": {
          "type": "string"
        }
      }
    },
...
    "List_Product_Response": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer"
        },
        "limit": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "docs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product_Response"
          }
        }
      }
    },
...
    "Category_Response": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "Brand_Response": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    }
  }
}